requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20.0
pillow
numpy
python-multipart